+++ ImpulsePlayer/src/main/kotlin/io/getimpulse/player/feature/cast/CastDisplay.kt
@@ -4,6 +4,7 @@ object CastDisplay {
 
     data class Route(
         val id: String,
+        val model: String?,
         val name: String,
     )
 }