+++ app/src/main/kotlin/io/getimpulse/example/model/Video.kt
@@ -2,5 +2,6 @@ package io.getimpulse.example.model
 
 data class Video(
     val title: String,
+    val subtitle: String,
     val url: String,
 )