+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-## 0.2.0
+### 0.2.1
 
-* [Picture In Picture] Fixed a case where Picture in Picture didn't recover when returning to the app. 
+* [Example] Adjusted example videos.
+* [Cast] Improved cast support.
+
+### 0.2.0
+
+* [Picture In Picture] Fixed a case where Picture in Picture didn't recover when returning to the app.
 * [Cast] Added cast support.
 * [Core] Adjusted core views and logic.
 
-## 0.1.0
+### 0.1.0
 
 * First version of the Impulse Player 🚀